Web: Files: refactoring: FilesRow, FilesTile renamed to RowsView, TilesView, pages moved to /src/
