+++ products/ASC.Files/Client/src/Files.jsx
@@ -12,9 +12,9 @@ import "./custom.scss";
 import i18n from "./i18n";
 import { I18nextProvider } from "react-i18next";
 import { regDesktop } from "@appserver/common/desktop";
-import Home from "./components/pages/Home";
-import Settings from "./components/pages/Settings";
-import VersionHistory from "./components/pages/VersionHistory";
+import Home from "./pages/Home";
+import Settings from "./pages/Settings";
+import VersionHistory from "./pages/VersionHistory";
 import ErrorBoundary from "@appserver/common/components/ErrorBoundary";
 import Panels from "./components/FilesPanels";
 import { AppServerConfig } from "@appserver/common/constants";
