+++ products/ASC.Files/Client/src/pages/Home/Section/Body/TilesView/BadgesFileTile.js
