+++ products/ASC.Files/Client/src/pages/Home/Section/Body/TilesView/FilesTileContent.js
