+++ products/ASC.Files/Client/src/pages/Home/Section/Body/TilesView/sub-components/Tile.js
