+++ products/ASC.Files/Client/src/pages/Home/Section/Paging/index.js
