+++ products/ASC.Files/Client/src/pages/Settings/Section/Body/ConnectedClouds.js
@@ -7,17 +7,17 @@ import Box from "@appserver/components/box";
 import Row from "@appserver/components/row";
 import RowContainer from "@appserver/components/row-container";
 import { withTranslation } from "react-i18next";
-import EmptyFolderContainer from "../../../../EmptyContainer/EmptyContainer";
-import BoxIcon from "../../../../../../public/images/icon_box.react.svg";
-import DropBoxIcon from "../../../../../../public/images/icon_dropbox.react.svg";
-import GoogleDriveIcon from "../../../../../../public/images/icon_google_drive.react.svg";
-import KDriveIcon from "../../../../../../public/images/icon_kdrive.react.svg";
-import NextCloudIcon from "../../../../../../public/images/icon_nextcloud.react.svg";
-import OneDriveIcon from "../../../../../../public/images/icon_onedrive.react.svg";
-import OwnCloudIcon from "../../../../../../public/images/icon_owncloud.react.svg";
-import SharePointIcon from "../../../../../../public/images/icon_sharepoint.react.svg";
-import WebDavIcon from "../../../../../../public/images/icon_webdav.react.svg";
-import YandexDiskIcon from "../../../../../../public/images/icon_yandex_disk.react.svg";
+import EmptyFolderContainer from "../../../../components/EmptyContainer/EmptyContainer";
+import BoxIcon from "../../../../../public/images/icon_box.react.svg";
+import DropBoxIcon from "../../../../../public/images/icon_dropbox.react.svg";
+import GoogleDriveIcon from "../../../../../public/images/icon_google_drive.react.svg";
+import KDriveIcon from "../../../../../public/images/icon_kdrive.react.svg";
+import NextCloudIcon from "../../../../../public/images/icon_nextcloud.react.svg";
+import OneDriveIcon from "../../../../../public/images/icon_onedrive.react.svg";
+import OwnCloudIcon from "../../../../../public/images/icon_owncloud.react.svg";
+import SharePointIcon from "../../../../../public/images/icon_sharepoint.react.svg";
+import WebDavIcon from "../../../../../public/images/icon_webdav.react.svg";
+import YandexDiskIcon from "../../../../../public/images/icon_yandex_disk.react.svg";
 import commonIconsStyles from "@appserver/components/utils/common-icons-style";
 import { inject, observer } from "mobx-react";
